requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Web-Scraping-Shopify-Image-Scraping/streamlit_app_techinstr.py: 4
Pillow == 9.4.0

# Web-Scraping-Shopify-Image-Scraping/streamlit_app_techinstr.py: 1
pandas == 1.5.3

# Web-Scraping-Shopify-Image-Scraping/streamlit_app_techinstr.py: 3
plotly == 5.15.0

# Web-Scraping-Shopify-Image-Scraping/streamlit_app_techinstr.py: 1
streamlit == 1.24.1
